Part 2 - Read from file
A new function is used to process the input file and get the data into an appropriate format. The main function gets an array with all the results (to manually check for any errors) and then prints the sum of the results, as requested. Refactored the calculate function to work with an array input - it would still work as expected if given sys.argv as a parameter